Add SetupModel and bind it to Setup page
A new view model called 'SetupModel' has been introduced and added to the 'Setup' page. The 'SetupModel' class contains methods for handling property changes and commands for navigating to the next step. The 'BindingContext' is also set for the 'AppShell' and 'Setup' classes. Two new packages 'CommunityToolkit.Maui' and 'CommunityToolkit.Mvvm' were added to the project.